backend/worker/requirements.txt
fastapi
uvicorn[standard]
redis
grpcio
protobuf==4.25.1
psycopg2-binary
aiofiles
etcd3
pynvml==11.5.0
# Optional heavy deps (comment in if you plan to use them)
# torch
# transformers
# diffusers
# vllm